[SWITCHYARD-41] - Moved soap tests to dir org.switchyard.component
